+++ ubo_app/constants.py
@@ -4,6 +4,8 @@ from __future__ import annotations
 import os
 from distutils.util import strtobool
 
+USERNAME = 'ubo'
+INSTALLATION_PATH = '/opt/ubo'
 DEBUG_MODE = strtobool(os.environ.get('UBO_DEBUG', 'False')) == 1
 SERVICES_PATH = (
     os.environ.get('UBO_SERVICES_PATH', '').split(':')