+++ ubo_app/store/main.py
@@ -14,33 +14,32 @@ from redux import (
     InitializationActionError,
     ReducerResult,
 )
-from ubo_gui.menu.types import SubMenuItem
+from ubo_gui.menu.types import Menu, SubMenuItem
 
-from ubo_app.store.app import (
-    RegisterAppAction,
-    RegisterRegularAppAction,
-)
-from ubo_app.store.keypad import (
+from ubo_app.store.app import RegisterAppAction, RegisterRegularAppAction
+from ubo_app.store.services.keypad import (
     Key,
     KeypadAction,
     KeypadEvent,
     KeypadKeyPressAction,
     KeypadKeyPressEvent,
 )
-from ubo_app.store.sound import (
-    SoundChangeVolumeAction,
-    SoundDevice,
-)
+from ubo_app.store.services.sound import SoundChangeVolumeAction, SoundDevice
 from ubo_app.store.status_icons import IconAction
+from ubo_app.store.update_manager_types import (
+    SetLatestVersionAction,
+    SetUpdateStatusAction,
+    VersionStatus,
+)
 
 if TYPE_CHECKING:
     from typing_extensions import TypeAlias
-    from ubo_gui.menu.types import Menu
 
 
 class MainState(Immutable):
     menu: Menu | None = None
     path: Sequence[int] = field(default_factory=list)
+    version: VersionStatus = field(default_factory=VersionStatus)
 
 
 class InitEvent(BaseEvent):
@@ -115,7 +114,7 @@ def main_reducer(  # noqa: C901
             msg = 'Main menu item is not a `SubMenuItem`'
             raise TypeError(msg)
 
-        main_menu_items = menu_items(main_menu_item.sub_menu)
+        main_menu_items = menu_items(cast(Menu, main_menu_item.sub_menu))
 
         desired_menu_item = main_menu_items[parent_index]
         if not isinstance(desired_menu_item, SubMenuItem):
@@ -128,20 +127,20 @@ def main_reducer(  # noqa: C901
             raise TypeError(msg)
 
         new_items = [
-            *cast(Sequence[Item], desired_menu_item.sub_menu.items),
+            *cast(Sequence[Item], cast(Menu, desired_menu_item.sub_menu).items),
             action.menu_item,
         ]
         desired_menu_item = replace(
             desired_menu_item,
             sub_menu=replace(
-                desired_menu_item.sub_menu,
+                cast(Menu, desired_menu_item.sub_menu),
                 items=new_items,
             ),
         )
         main_menu_item = replace(
             main_menu_item,
             sub_menu=replace(
-                main_menu_item.sub_menu,
+                cast(Menu, main_menu_item.sub_menu),
                 items=[
                     desired_menu_item if index == parent_index else item
                     for index, item in enumerate(main_menu_items)
@@ -162,5 +161,23 @@ def main_reducer(  # noqa: C901
 
     if isinstance(action, SetMenuPathAction):
         return replace(state, path=action.path)
+
+    if isinstance(action, SetLatestVersionAction):
+        return replace(
+            state,
+            version=replace(
+                state.version,
+                latest_version=action.latest_version,
+            ),
+        )
+
+    if isinstance(action, SetUpdateStatusAction):
+        return replace(
+            state,
+            version=replace(
+                state.version,
+                update_status=action.status,
+            ),
+        )
 
     return state